First game area - Big environment update
Finished creating the rest of the background assets for the game, and set up the background for the first playable area. ALL needed prefabs are in the Prefabs > Environment folder. Dragging all the prefabs there into the hierarchy is all that's needed to show the environment in the right context.